+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Now the Mastodon related classes and objects are abstracted into `pyxavi` ([#42](https://github.com/XaviArnaus/janitor/pull/42))
+- Now the Publish class is abstracted into `pyxavi` ([#44](https://github.com/XaviArnaus/janitor/pull/44))
 
 ## [v0.5.2](https://github.com/XaviArnaus/janitor/releases/tag/v0.5.2)